% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_up.R
\name{signif_up}
\alias{signif_up}
\title{round up to most significant digit}
\usage{
signif_up(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric
}
\description{
\emph{round up} a number to the most meaningful digits per scale of the number
}
\examples{
signif_up(1320.65) # 1400
signif_up(132.5) # 140
signif_up(32.6)  # 40
signif_up(3.6)   # 4
signif_up(0.0654)  # 0.07
signif_up(c(1320.654, 3675.675, 5.7, 0.76))
}
