% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_vr_by_grp_fn.R
\name{summarise_vr_by_grp_fn}
\alias{summarise_vr_by_grp_fn}
\title{group by and summarise with multiple functions}
\usage{
summarise_vr_by_grp_fn(
  dta,
  grp = vars(),
  vr = vars(),
  fn = list(mean = mean, median = median),
  ...
)
}
\arguments{
\item{dta}{input dataset}

\item{grp}{group by vars defined by vars()}

\item{vr}{variable to be summarised defined by vars()}

\item{fn}{a named list of functions}

\item{...}{any other input listed fn functions}
}
\value{
grouped and summarised dataset
}
\description{
group by and summarise with multiple functions
}
\examples{
summarise_vr_by_grp_fn(
dta = iris,
grp = vars(Species),
vr = vars("Sepal.Length", "Petal.Width"),
fn = list(avg = mean, median = median)
)
summarise_vr_by_grp_fn(
dta = iris,
grp = vars(Species),
vr = vars(Sepal.Length, Petal.Width),
fn = list(avg = mean, median = median)
)
}
